added more DTO , mappers and Post request to create a stock
when you make a model, you have to create a dto

each dto potentially will be specialized for specific task, hiding unnecessary info

make a mapper to glue both DTO and the original class. In the end Both DTO and class are just objects!

issue comes in when mapping both and passing to the DB back and forth

lastly you have to be crafty when it comes to handle DTO to the database check the StockController.cs class Create request for more insight